requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Jinja2==3.1.2
numpy==1.26.0
pandas==2.1.1
pyodbc==4.0.39
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9